requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.38
langchain-google-genai>=1.0.0
langchain-neo4j>=0.0.3


# Data processing
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0

# Vector stores and embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0


# Graph database
neo4j>=5.0.0

# NLP and Intent classification
spacy>=3.4.0



# Environment management
python-dotenv>=1.0.0



# Additional dependencies for production
requests>=2.28.0
typing-extensions>=4.0.0

# Guardrails and PII protection(may cause issues please refer documentation)
guardrails-ai>=0.6.0
presidio-anonymizer>=2.2.0
presidio-analyzer>=2.2.0